feat(mixed): code adding
adding of part of code to add collision for power up + preparation power up states

+++ SkySurfer/Assets/Scripts/Entities/PlayerEntity/Player.cs
@@ -31,7 +31,7 @@ namespace SkySurfer.Assets.Scripts.Entities.PlayerEntity
             _powerUpTime = time;
         }
 
-        public float GetHp()
+        public int GetHp()
         {
             return _hp;
         }
